\name{r_dist_sim}
\alias{r_dist_sim}
\title{Méthode de rejet}
\usage{
r_dist_sim (f, n, a)
}
\description{
Fonction de simulation de densité de la distance R par méthode de rejet.

Input

f : Fonction de densité

n : Nombre d'observations

a : Borne supérieure de f

Output

r : Densité estimée
}

